// backend/src/modules/data-models/interfaces/field-type.enum.ts
export enum FieldType {
  STRING = 'string',
  TEXT = 'text',
  RICH_TEXT = 'richText',
  INTEGER = 'integer',
  FLOAT = 'float',
  DECIMAL = 'decimal',
  BOOLEAN = 'boolean',
  DATE = 'date',
  DATETIME = 'datetime',
  TIME = 'time',
  ENUM = 'enum',
  JSON = 'json',
  ARRAY = 'array',
  REFERENCE = 'reference',
  FILE = 'file',
  IMAGE = 'image',
  EMAIL = 'email',
  URL = 'url',
  PHONE = 'phone',
  COLOR = 'color',
  GEO = 'geo',
}